requirements.txt
queuelib
threaded
pyyaml
certifi
